backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.7.1
PyJWT==2.9.0
pytest==8.3.3
GitPython==3.1.43
